+++ packages/db/src/query/device.ts
@@ -1,8 +1,9 @@
 import { type ExtractTablesWithRelations, type SQLWrapper, and, eq, ilike, like, sql } from "drizzle-orm";
 import type { MySqlTransaction } from "drizzle-orm/mysql-core";
 import type { PlanetScalePreparedQueryHKT, PlanetscaleQueryResultHKT } from "drizzle-orm/planetscale-serverless";
+import { type MathNumericType, type Matrix, all, create } from "mathjs";
 import db from "../";
-import { device, deviceHistory } from "../schema";
+import { device, deviceHistory, deviceToPeak, sensorDataSequence, userData } from "../schema";
 
 export async function getDevicesByUser(userId: string, search?: string) {
     const conditions: SQLWrapper[] = [eq(device.userId, userId)];
@@ -54,6 +55,80 @@ export async function deleteDevice(id: number, userId: string) {
 
         await copyToHistoryTable(trx, deviceToDelete);
         await trx.delete(device).where(eq(device.id, id));
+    });
+}
+
+export async function updatePowerOfDevices(userId: string) {
+    return db.transaction(async (trx) => {
+        const devicesWithPeaks = await trx
+            .select()
+            .from(device)
+            .leftJoin(deviceToPeak, eq(deviceToPeak.deviceId, device.id))
+            .leftJoin(sensorDataSequence, eq(sensorDataSequence.id, deviceToPeak.sensorDataSequenceId))
+            .where(eq(device.userId, userId));
+
+        const devices = Array.from(new Set(devicesWithPeaks.map((device) => device.device.id)));
+        const flattenPeak = devicesWithPeaks
+            .filter((x) => x.device_to_peak && x.sensor_data_sequence)
+            .map((device) => ({
+                sequenceId: device.sensor_data_sequence?.id ?? "",
+                power: device.sensor_data_sequence?.averagePeakPower ?? 0,
+                device: device.device_to_peak?.deviceId ?? 0,
+            }));
+
+        const peaks = Object.values(
+            flattenPeak.reduce(
+                (acc, obj) => {
+                    if (!acc[obj.sequenceId]) {
+                        acc[obj.sequenceId] = { sequence: obj.sequenceId, devices: [], power: obj.power };
+                    }
+                    acc[obj.sequenceId].devices.push(obj.device);
+                    return acc;
+                },
+                {} as { [key: string]: { sequence: string; devices: number[]; power: number } },
+            ),
+        );
+
+        const math = create(all, {});
+
+        // Extract matrices
+        const A = math.matrix(
+            peaks.map((dp) => {
+                const row = devices.map((device) => (dp.devices.includes(device) ? 1 : 0));
+                return row;
+            }),
+        );
+        const b = math.matrix(peaks.map((dp) => [dp.power]));
+
+        // Solve the least squares problem: (A^T A) x = A^T b
+        const At = math.transpose(A);
+        const AtA = math.multiply(At, A);
+        const Atb = math.multiply(At, b);
+        const solution = math.lusolve(AtA, Atb);
+
+        // Extract the solution values and save them to the database
+        const result = solution
+            .toArray()
+            .map((value, index) => ({ [devices[index]]: (value as MathNumericType[])[0] }));
+        for (const deviceId of devices) {
+            const powerEstimationRaw = result.find((r) => r[deviceId])?.[deviceId];
+            const powerEstimation = powerEstimationRaw ? Number(powerEstimationRaw) : null;
+            const correctedPowerEstimation = powerEstimation && powerEstimation >= 0 ? powerEstimation : null; // power needs to be greater than 0.
+            await trx.update(device).set({ powerEstimation: correctedPowerEstimation }).where(eq(device.id, deviceId));
+        }
+
+        // Calculate R^2
+        const residuals = math.subtract(b, math.multiply(A, solution));
+        const subtraction = math.subtract(b, math.mean(b)) as Matrix;
+        const SST = math.sum(math.map(subtraction, math.square));
+        const SSR = math.sum(math.map(residuals, math.square));
+        let rSquared: number;
+        if (SST === 0) {
+            rSquared = SSR === 0 ? 1 : 0;
+        } else {
+            rSquared = math.subtract(1, math.divide(SSR, SST)) as number;
+        }
+        await trx.update(userData).set({ devicePowerEstimationRSquared: rSquared }).where(eq(userData.userId, userId));
     });
 }
