+++ packages/db/src/query/sensor.ts
@@ -1,23 +1,11 @@
 import { AggregationType, UserHasSensorOfSameType } from "@energyleaf/lib";
 import { SensorAlreadyExistsError } from "@energyleaf/lib/errors/sensor";
-import { and, asc, between, desc, eq, gte, inArray, isNotNull, lt, lte, ne, or, sql } from "drizzle-orm";
-import { type MathNumericType, type Matrix, all, create } from "mathjs";
+import { and, between, desc, eq, gte, isNotNull, lt, lte, ne, or, sql } from "drizzle-orm";
 import { nanoid } from "nanoid";
 import db from "../";
-import {
-    device,
-    deviceToPeak,
-    sensor,
-    sensorData,
-    sensorDataSequence,
-    sensorHistory,
-    sensorToken,
-    user,
-    userData,
-} from "../schema";
+import { sensor, sensorData, sensorHistory, sensorToken, user, userData } from "../schema";
 import {
     type SensorDataSelectType,
-    type SensorDataSequenceType,
     type SensorInsertType,
     type SensorSelectTypeWithUser,
     SensorType,
@@ -40,239 +28,6 @@ export async function getAllSensors(active?: boolean) {
     }
 
     return db.select().from(sensor);
-}
-
-function calculateMedian(values: SensorDataSelectType[]) {
-    const sorted = [...values].sort((a, b) => a.value - b.value);
-    const middle = Math.floor(sorted.length / 2);
-    return sorted.length % 2 !== 0 ? sorted[middle].value : (sorted[middle - 1].value + sorted[middle].value) / 2;
-}
-
-function calculateMAD(values: SensorDataSelectType[], scale = 1.4826, medValue?: number) {
-    let medianValue: number | undefined = medValue;
-    if (!medianValue) {
-        medianValue = calculateMedian(values);
-    }
-    const deviations = values.map((value) => ({
-        ...value,
-        value: Math.abs(value.value - medianValue),
-    }));
-    return scale * calculateMedian(deviations);
-}
-
-/**
- * Returns average power in Watt.
- */
-function calculateAveragePower(sensorData: SensorDataSelectType[]) {
-    if (sensorData.length < 2) {
-        throw new Error("Cant calculate average power for one or no entry.");
-        //Won't happen, because our peaks have a minimum length
-    }
-
-    const powerSum = sensorData.reduce((acc, curr, index) => {
-        if (index === 0) {
-            return acc;
-        }
-
-        const timeDiffInHours = (curr.timestamp.getTime() - sensorData[index - 1].timestamp.getTime()) / 1000 / 60 / 60;
-        return acc + (curr.value / timeDiffInHours) * 1000; // Add power in Watt
-    }, 0);
-
-    return powerSum / sensorData.length;
-}
-
-interface Sequence {
-    start: Date;
-    end: Date;
-    isAtStart: boolean;
-    averagePowerIncludingBaseLoad: number;
-}
-
-function findSequences(values: SensorDataSelectType[], threshold: number) {
-    const sequences: Sequence[] = [];
-    let i = 0;
-
-    while (i < values.length) {
-        const entry = values[i];
-
-        if (entry.value > threshold) {
-            const isStart = i === 0;
-            let sequenceEnd = i + 1;
-
-            while (sequenceEnd < values.length && values[sequenceEnd].value > threshold) {
-                sequenceEnd++;
-            }
-
-            const sequenceLength = sequenceEnd - i;
-
-            // only mark as peak if longer then 2min and not marked as anomaly yet
-            if (sequenceLength > 5) {
-                const avgPeakPower = calculateAveragePower(values.slice(i, sequenceEnd));
-
-                sequences.push({
-                    start: entry.timestamp,
-                    end: values[sequenceEnd - 1].timestamp,
-                    isAtStart: isStart,
-                    averagePowerIncludingBaseLoad: avgPeakPower,
-                });
-            }
-            i = sequenceEnd;
-        } else {
-            i++;
-        }
-    }
-
-    return sequences;
-}
-
-/**
- * Finds peaks in a given set of values
- * for peaks 5 is the default threshold, for anomalies 1000 is a good fit
- */
-export function findPeaks(
-    thresholdValues: SensorDataSelectType[],
-    consideredValues: SensorDataSelectType[],
-    threshold = 5,
-    medianValue?: number,
-) {
-    let median: number | undefined = medianValue;
-    if (!median) {
-        median = calculateMedian(thresholdValues);
-    }
-    const madValue = calculateMAD(thresholdValues, 1.4826, median);
-    const processedValues = consideredValues.map((value) => ({
-        ...value,
-        value: Math.abs(value.value - median),
-    }));
-    return findSequences(processedValues, madValue * threshold);
-}
-
-interface FindAndMarkPeaksProps {
-    sensorId: string;
-    start: Date;
-    end: Date;
-    type: "peak" | "anomaly";
-}
-
-export async function findAndMark(props: FindAndMarkPeaksProps, threshold = 5) {
-    const { sensorId, start, end, type } = props;
-
-    // we shift the start 12 hours back, so we have a bigger sample for the threshold
-    const sequenceStart = new Date(start);
-    sequenceStart.setHours(sequenceStart.getHours() - 24, 0, 0, 0);
-
-    try {
-        return await db.transaction(async (trx) => {
-            const calcDbData = await trx
-                .select()
-                .from(sensorData)
-                .where(and(eq(sensorData.sensorId, sensorId), between(sensorData.timestamp, sequenceStart, end)))
-                .orderBy(asc(sensorData.timestamp));
-
-            // make sure we have at least 3 hours of reference data
-            if (calcDbData.length === 0 || calcDbData.length < 720) {
-                return [];
-            }
-
-            const calcData = calcDbData
-                .map((d, i) => {
-                    return {
-                        ...d,
-                        value: i === 0 ? 0 : Number(d.value) - Number(calcDbData[i - 1].value),
-                    };
-                })
-                .slice(1);
-
-            // check if all values are integers if so we know that the sensor has no pin
-            if (calcData.every((d) => Number.isInteger(d.value))) {
-                return [];
-            }
-
-            const energyData = calcData.filter((d) => {
-                return d.timestamp.getTime() >= start.getTime();
-            });
-
-            const sequencesInConsideredPeriod = findPeaks(calcData, energyData, threshold);
-
-            const sequencesBeforeConsideredPeriod = await trx
-                .select()
-                .from(sensorDataSequence)
-                .where(and(eq(sensorDataSequence.sensorId, sensorId), between(sensorDataSequence.start, start, end)));
-
-            const peaksInDay = sequencesInConsideredPeriod
-                .map((d) => ({ start: d.start, end: d.end }))
-                .concat(sequencesBeforeConsideredPeriod.map((d) => d));
-            const calcDataWithoutPeaks = calcData.filter(
-                (d) => !peaksInDay.some((peak) => d.timestamp >= peak.start && d.timestamp <= peak.end),
-            );
-            const baseLoad = calculateAveragePower(calcDataWithoutPeaks);
-            let peaks: SensorDataSequenceType[] = sequencesInConsideredPeriod.map((d) => ({
-                ...d,
-                id: nanoid(30),
-                averagePeakPower: d.averagePowerIncludingBaseLoad - baseLoad,
-                type,
-                sensorId,
-            }));
-
-            if (props.type === "anomaly") {
-                // if it is anomaly make sure there at least 30min apart from previous ones to avoid double marking
-                const lastAnomaly = sequencesBeforeConsideredPeriod.find((d) => d.type === "anomaly");
-                if (lastAnomaly) {
-                    peaks = peaks.filter((d) => d.start.getTime() - lastAnomaly.end.getTime() > 30 * 60 * 1000);
-                }
-            }
-            if (peaks.length !== 0) {
-                // check before value if this peak is part of another peak sequence
-                // if so we don't mark it as peak
-                if (sequencesInConsideredPeriod[0].isAtStart) {
-                    const lastSequenceOfSensorQuery = await trx
-                        .select()
-                        .from(sensorDataSequence)
-                        .where(
-                            and(
-                                eq(sensorDataSequence.sensorId, sensorId),
-                                lt(sensorDataSequence.end, start),
-                                eq(sensorDataSequence.type, type),
-                            ),
-                        )
-                        .orderBy(desc(sensorDataSequence.end))
-                        .limit(1);
-                    const lastSequenceOfSensor =
-                        lastSequenceOfSensorQuery.length > 0 ? lastSequenceOfSensorQuery[0] : null;
-
-                    const beforeIndex = calcData.findIndex((d) => d.id === energyData[0].id);
-                    const lastEntryBeforeInterval = calcData[beforeIndex - 1];
-
-                    if (
-                        lastSequenceOfSensor &&
-                        lastSequenceOfSensor.end.getTime() >= lastEntryBeforeInterval.timestamp.getTime()
-                    ) {
-                        await trx
-                            .update(sensorDataSequence)
-                            .set({ end: sequencesInConsideredPeriod[0].end })
-                            .where(eq(sensorDataSequence.id, lastSequenceOfSensor.id));
-                    } else {
-                        await trx.insert(sensorDataSequence).values(peaks);
-                    }
-                } else {
-                    await trx.insert(sensorDataSequence).values(peaks);
-                }
-            }
-
-            // if there is an anomaly in the last 24 hours return nothing to avoid sending another mail
-            if (props.type === "anomaly") {
-                if (sequencesInConsideredPeriod.length !== 0) {
-                    if (sequencesBeforeConsideredPeriod.some((d) => d.type === "anomaly")) {
-                        return [];
-                    }
-                }
-            }
-
-            return peaks;
-        });
-    } catch (err) {
-        return [];
-    }
 }
 
 export async function getEnergyForSensorInRange(
@@ -354,8 +109,6 @@ export async function getEnergyForSensorInRange(
         valueOut: row.valueOut ? (index === 0 ? 0 : Number(row.valueOut) - Number(query[index - 1].valueOut)) : null,
         valueCurrent: row.valueCurrent,
         timestamp: new Date(row.timestamp),
-        isPeak: false,
-        isAnomaly: false,
     }));
 
     return results.slice(1);
@@ -526,111 +279,6 @@ export async function getAvgEnergyConsumptionForUserInComparison(userId: string)
     });
 
     return query;
-}
-
-/**
- *  adds or updates a peak in the database
- */
-export async function updateDevicesForPeak(sensorDataSequenceId: string, deviceIds: number[]) {
-    return db.transaction(async (trx) => {
-        await trx.delete(deviceToPeak).where(eq(deviceToPeak.sensorDataSequenceId, sensorDataSequenceId));
-
-        for (const deviceId of deviceIds) {
-            await trx.insert(deviceToPeak).values({
-                deviceId,
-                sensorDataSequenceId,
-            });
-        }
-    });
-}
-
-export async function updatePowerOfDevices(userId: string) {
-    return db.transaction(async (trx) => {
-        const devicesWithPeaks = await trx
-            .select()
-            .from(device)
-            .leftJoin(deviceToPeak, eq(deviceToPeak.deviceId, device.id))
-            .leftJoin(sensorDataSequence, eq(sensorDataSequence.id, deviceToPeak.sensorDataSequenceId))
-            .where(eq(device.userId, userId));
-
-        const devices = Array.from(new Set(devicesWithPeaks.map((device) => device.device.id)));
-        const flattenPeak = devicesWithPeaks
-            .filter((x) => x.device_to_peak && x.sensor_data_sequence)
-            .map((device) => ({
-                sequenceId: device.sensor_data_sequence?.id ?? "",
-                power: device.sensor_data_sequence?.averagePeakPower ?? 0,
-                device: device.device_to_peak?.deviceId ?? 0,
-            }));
-
-        const peaks = Object.values(
-            flattenPeak.reduce(
-                (acc, obj) => {
-                    if (!acc[obj.sequenceId]) {
-                        acc[obj.sequenceId] = { sequence: obj.sequenceId, devices: [], power: obj.power };
-                    }
-                    acc[obj.sequenceId].devices.push(obj.device);
-                    return acc;
-                },
-                {} as { [key: string]: { sequence: string; devices: number[]; power: number } },
-            ),
-        );
-
-        const math = create(all, {});
-
-        // Extract matrices
-        const A = math.matrix(
-            peaks.map((dp) => {
-                const row = devices.map((device) => (dp.devices.includes(device) ? 1 : 0));
-                return row;
-            }),
-        );
-        const b = math.matrix(peaks.map((dp) => [dp.power]));
-
-        // Solve the least squares problem: (A^T A) x = A^T b
-        const At = math.transpose(A);
-        const AtA = math.multiply(At, A);
-        const Atb = math.multiply(At, b);
-        const solution = math.lusolve(AtA, Atb);
-
-        // Extract the solution values and save them to the database
-        const result = solution
-            .toArray()
-            .map((value, index) => ({ [devices[index]]: (value as MathNumericType[])[0] }));
-        for (const deviceId of devices) {
-            const powerEstimationRaw = result.find((r) => r[deviceId])?.[deviceId];
-            const powerEstimation = powerEstimationRaw ? Number(powerEstimationRaw) : null;
-            const correctedPowerEstimation = powerEstimation && powerEstimation >= 0 ? powerEstimation : null; // power needs to be greater than 0.
-            await trx.update(device).set({ powerEstimation: correctedPowerEstimation }).where(eq(device.id, deviceId));
-        }
-
-        // Calculate R^2
-        const residuals = math.subtract(b, math.multiply(A, solution));
-        const subtraction = math.subtract(b, math.mean(b)) as Matrix;
-        const SST = math.sum(math.map(subtraction, math.square));
-        const SSR = math.sum(math.map(residuals, math.square));
-        let rSquared: number;
-        if (SST === 0) {
-            rSquared = SSR === 0 ? 1 : 0;
-        } else {
-            rSquared = math.subtract(1, math.divide(SSR, SST)) as number;
-        }
-        await trx.update(userData).set({ devicePowerEstimationRSquared: rSquared }).where(eq(userData.userId, userId));
-    });
-}
-
-export async function getSequencesBySensor(sensorId: string) {
-    return db.select().from(sensorDataSequence).where(eq(sensorDataSequence.sensorId, sensorId));
-}
-
-export async function getDevicesByPeak(sensorDataSequenceId: string) {
-    return db
-        .select({
-            id: deviceToPeak.deviceId,
-            name: device.name,
-        })
-        .from(deviceToPeak)
-        .innerJoin(device, eq(device.id, deviceToPeak.deviceId))
-        .where(eq(deviceToPeak.sensorDataSequenceId, sensorDataSequenceId));
 }
 
 /**
